Removed byte cast

+++ Pinta.Core/Effects/ColorBgra.cs
@@ -347,7 +347,7 @@ namespace Pinta.Core
 
             ret.B = ClampToByte(Lerp(from.B, to.B, frac));
             ret.G = (byte)ClampToByte(Lerp(from.G, to.G, frac));
-            ret.R = (byte)ClampToByte(Lerp(from.R, to.R, frac));
+            ret.R = ClampToByte(Lerp(from.R, to.R, frac));
             ret.A = ClampToByte(Lerp(from.A, to.A, frac));
 
             return ret;
